Updated the Admin and Organizer Middleware
Try logging in as Admin: They should have access to /admin/dashboard and /organizer/dashboard.
Try logging in as Organizer: They should have access to /organizer/dashboard but not /admin/dashboard.
Try logging in as a User: They should be redirected to /dashboard(this is gonna be our Home page)